telecomsteve/requirements.txt
attrs==21.4.0
cachetools==5.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.26.0
requests-cache==0.7.4
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
url-normalize==1.4.3
urllib3==1.26.6
Werkzeug==2.0.1
